pyproject.toml
[tool.poetry]
name = "py-progress-tracker"
version = "0.4.0"
description = "A simple benchmarking library"
authors = ["Zama <hello@zama.ai>"]
packages = [
    { include = "py_progress_tracker" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
colorama = "^0.4.4"
psutil = "^5.8.0"
py-cpuinfo = "^8.0.0"
tabulate = "^0.8.9"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
